Reworked how the developers are listed.
Per legal counsel advice, anyone that has contributed more than a 'de minimus'
contribution can now be listed as a principal developer.  The lead developer
is the person responsible for pushing the form through the complete process.
Given this change, it didn't make sense to list all principal developers on
the form when they can be listed on the abstract instead.  This change
incorporates that realization.