Improved sample abstracts, and modified fake email addresses.
The original email address pointed to gmail.com.  Since this is a real domain,
it would have been possible for someone to register the email address and
receive mail from the sample abstracts.  This modification points to fake
domains, or domains controlled by the US government.

This modification also improves the text of the sample abstracts, making them
more readable.