Minor cleanup in preparation for the complete 2.0 rewrite.
After more than a year of testing, I've discovered that the 1.0 policy
requires a rewrite to make it more easily used by ARL's workforce, and to make
it easier for others to adapt for their own needs.  This commit is not that
rewrite; it is just in preparation for the rewrite.  The rewrite will be done
on a separate branch dedicated to that purpose.