Rewrote opening of README.md.
README.md is now going to be a landing page that gives a high-level overview
and some extra information about the policy.  The instructions are in their
own file, as are the goals and rationale.  This makes it easier for users to
find the section that they are interested in, skipping the other portions.