Added the latest draft of the ARL Major Release Form.
This is the latest draft of the ARL Major Release Form.  This draft is not yet
an official ARL form and cannot be used for processing paperwork.  As a note,
this form cannot be turned into an official form until the 2.0 policy is
signed by the Director of ARL.  Until and unless that happens, this form is
no better than the mockup form.